requirements.txt
aiohttp==3.7.4.post0
astroid==2.5.6
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
discord.py==1.7.2
Flask==2.0.1
greenlet==1.1.0
idna==2.10
isort==5.8.0
itsdangerous==2.0.1
Jinja2==3.0.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
msgpack==1.0.2
multidict==5.1.0
Pillow==8.2.0
pylint==2.8.3
pynvim==0.4.3
python-dotenv==0.17.1
requests==2.25.1
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.5
Werkzeug==2.0.1
wrapt==1.12.1
yapf==0.31.0
yarl==1.6.3
